restructure to src/ for source, tests/ for tests, implement tests for naming convention update using pytest, change Tk to pyQT in SB (faster), normalise train naming convention to newer translink version in SB, separate variable for revenue only trains, update .gitignore to be conventional for python codebases
