+++ conftest.py
@@ -1,0 +1,5 @@
+import sys
+import os
+
+# Add the 'src' directory to the path so internal imports work
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), 'src')))
