+++ src/HASTUS_Converter.py
@@ -12,8 +12,7 @@ from tkinter.filedialog import askopenfilename
 
 import traceback
 import logging
-
-
+import MTP_constants
 
 
 
@@ -33,14 +32,10 @@ hastuscopyfile = True if os.path.basename(__file__) == 'HASTUS_Converter - Copy.
 # --------------------------------------------------------------------------------------------------- #
 
 
-
-
-
 ### Dictionary used to title reports for each daycode
 daycode_dict = {'120':'muwt','4':'f','2':'a','1':'s'}
 
-### Dictionary used for printouts when running the report
-weekdaykey_dict = {'120':'Mon-Thu','64': 'Mon','32': 'Tue','16': 'Wed','8':  'Thu', '4':  'Fri','2':  'Sat','1':  'Sun'}
+### Dictionary used for printouts when running the report - const moved
 
 ### Do not want these entries included at all in the HASTUSExport
 entries_to_exclude = ['RSWJ','YNA','RSF',
@@ -90,26 +85,6 @@ HASTUS_stableconverter = {
     }
 
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 def TTS_H(path, mypath = None):
     
     copyfile = '\\'.join(path.split('/')[0:-1]) != mypath and mypath is not None
@@ -142,16 +117,6 @@ def TTS_H(path, mypath = None):
             print(myhastuspath)
             print('———————————————————————————————————————————————————————————————————————————————\n')
         
-        
-    
-        
-        
-        
-        
-        
-        
-        
-        
       
         tree = ET.parse(filename)
         root = tree.getroot()
@@ -173,18 +138,9 @@ def TTS_H(path, mypath = None):
                 
         if tn_doubles:
             print('           Error: Duplicate train numbers')
-            for tn,day in tn_doubles: print(f' - 2 trains runnnig on {weekdaykey_dict.get(day)} with train number {tn} - ')
+            for tn,day in tn_doubles: print(f' - 2 trains runnnig on {MTP_constants.WEEKDAYKEY.get(day)} with train number {tn} - ')
             time.sleep(15)
             sys.exit() 
-        
-        
-        
-        
-        
-        
-        
-        
-        
         
         
         def format_run(run):
@@ -278,23 +234,6 @@ def TTS_H(path, mypath = None):
     
         SORT_ORDER_WEEK = ['1','2','4','120']
         d_list.sort(key=SORT_ORDER_WEEK.index)
-        
-                        
-                        
-                        
-                        
-                        
-                        
-                        
-                        
-                        
-        
-                    
-        
-    
-              
-                
-        
         
         
         ### uniquestations_dict, network_vrt_dict and virtual run time (vrt) dictionaries for each line are used to determine direction (Up or Down)
@@ -658,10 +597,6 @@ def TTS_H(path, mypath = None):
             }
         
         
-        
-        
-        
-        
         def create_textfile(weekdaykey):
             """ Creates a HASTUS Export textfile for a single day of operations """
             
@@ -981,9 +916,6 @@ def TTS_H(path, mypath = None):
     #                 print(f'{line} - {firstonline} then {secondonline}\n')
     # =============================================================================
                     
-                
-                
-                
     
                 tripinfo = [tn,empt,drct,stations]
                 
@@ -1035,11 +967,6 @@ def TTS_H(path, mypath = None):
                             stop = station[3]
                             
                             
-                            
-                            
-                            
-                            
-                            
                             hhmmss = hhmm + ':00'
                             stationtosignal = str(pd.Timedelta(hhmmss) + pd.Timedelta(seconds=60))
                             signaltostation = str(pd.Timedelta(hhmmss) - pd.Timedelta(seconds=60))
@@ -1069,17 +996,11 @@ def TTS_H(path, mypath = None):
                                 
                             
                             
-                            
-       
-                            
                             ### Write the station
                             wl([nl,'triptp',l,sID,l,hhmm,l,zero,l,stop,l,f'{run}_{tn}'])
                             
                             
     
-                            
-    
-                            
                             
                             ### Input EJ28 Signal Turnback after the last station to avoid mismatched plaform errors
                             if station == stations[-1]:
@@ -1092,14 +1013,9 @@ def TTS_H(path, mypath = None):
                                     #time + 1 minutes
                    
                             
-                   
-                    
-                   
-                    
-                   
                     
                 o.close()
-                print(f'All trains on {weekdaykey_dict.get(weekdaykey)} have been processed')
+                print(f'All trains on {MTP_constants.WEEKDAYKEY.get(weekdaykey)} have been processed')
                 # print('—————————————————————————————————————————————————————')
                 print('\n\n')
                 if hastuscopyfile:
@@ -1115,17 +1031,9 @@ def TTS_H(path, mypath = None):
         ### Run the create_textfile function for every day present in the rsx
         for day in d_list:
             # print('—————————————————————————————————————————————————————')
-            print(f'Processing trains on {weekdaykey_dict.get(day)}...', end='\r')
+            print(f'Processing trains on {MTP_constants.WEEKDAYKEY.get(day)}...', end='\r')
             create_textfile(day)
             
-                
-    
-        
-        
-        
-        
-        
-        
         
         if __name__ == "__main__":
             print(f'(runtime: {time.time()-start_time:.2f}seconds)')
