+++ src/MTP_constants.py
@@ -27,12 +27,12 @@ WEEKDAY_KEYS_MASTER = {
 
 DAY_PRIORITY = ['64','32','16','8','4','2','1','120']
 SORT_ORDER_WEEK = ['64','32','16','8','120','4','2','1']
-SORT_ORDER_UNIT = ['REP','NGR','NGRE','IMU100','EMU','SMU','HYBRID','ICE','DEPT']
+#SORT_ORDER_UNIT = ['REP','NGR','NGRE','IMU100','EMU','SMU','HYBRID','ICE','DEPT']
 ID_TO_SHORT = {k: v['short'] for k, v in WEEKDAY_KEYS_MASTER.items()}
 ID_TO_LONG  = {k: v['long'] for k, v in WEEKDAY_KEYS_MASTER.items()}
 ID_TO_ALIAS = {k: v['alias'] for k, v in WEEKDAY_KEYS_MASTER.items()}
 
-# Universal Reverse Map (maps ANY name/alias back to the ID)
+# Universal Reverse Map (maps name/alias back to the ID)
 NAME_TO_ID = {}
 for uid, info in WEEKDAY_KEYS_MASTER.items():
     for val in info.values():
@@ -40,7 +40,7 @@ for uid, info in WEEKDAY_KEYS_MASTER.items():
 
 
 SORT_ORDER_WEEK = ['64','32','16','8','120','4','2','1'] 
-SORT_ORDER_UNIT = ['REP','NGR','NGRE','IMU100','EMU','SMU','HYBRID', 'DEPT']
+SORT_ORDER_UNIT = ['QMU', 'REP','NGR','NGRE','IMU100','EMU','SMU','HYBRID', 'DEPT']
 
 ### location stuff 
 # if new location found, update locations , flag it 
@@ -48,7 +48,7 @@ SORT_ORDER_UNIT = ['REP','NGR','NGRE','IMU100','EMU','SMU','HYBRID', 'DEPT']
 # print new location in excel file for easy debugging 
 
     
-
+# update stabling yard locations here 
 YARDS = {
     'Wulkuraka':    ['WFE','WFW','FEE'],
     'Ipswich':      ['IPSS','IPS'],
@@ -73,11 +73,55 @@ YARDS = {
 }
 
 
+# update internal list of non stable yards 
 NON_STABLE_LOCATIONS = ['IPS','MNY','CAB','NBR','GYN','RS','BHI']
 
 
 
+### STYLING (colours)
+
+ALERT = "#CC194C"
+GREY  = "#CCCCCC"
+REP   = "#FFB7B7"
+NGR   = "#E4DFEC"
+NGRE  = "#FFFF93"
+IMU   = "#FDE9D9"
+EMU   = "#DAEEF3"
+SMU   = "#F2DCDB"
+DEPT  = "#EBF1DE"
+QMU = "#A7B48D"
 
 
+UNBALANCED_YELLOW = "#CCB233"
+WHITE = "#FFFFFF"
+
+# Map used by the builder to generate families
+FAMILY_BG = {
+    "REP":    REP,
+    "NGR":    NGR,
+    "NGRE":   NGRE,
+    "IMU100": IMU,
+    "EMU":    EMU,
+    "HYBRID": EMU,   # shared EMU palette
+    "SMU":    SMU,
+    "DEPT":   DEPT,
+    "QMU":    QMU
 
 
+}
+
+
+# Explicit override
+TRAIN_TYPE_MASK = {
+    'empty_6-rep': 'Empty_6-QMU',
+    '6-rep': '6-QMU',
+    '6-qmu_(aw0)_surface': 'Empty_6-QMU',
+    '6-qmu_(aw3)_surface': '6-QMU',
+    '6-ngr_(aw0)_surface': 'Empty_6-NGR',
+    '6-ngr_(aw3)_surface': '6-NGR',
+    'qmu_s': '6-QMU',
+    'empty_qmu_s': 'Empty_6-QMU',
+    '6-ngr_s': '6-NGR',
+    'empty_6-ngr_s': 'Empty_6-NGR',
+    'ngr_s': '6-NGR',
+}
