+++ src/debug.py
@@ -1,0 +1,5 @@
+import pkg_resources
+
+installed_packages = pkg_resources.working_set
+for package in installed_packages:
+    print(f"{package.key}=={package.version}")
