+++ src/utils.py
@@ -24,3 +24,4 @@ def csl(string):
         if x not in output:
             output.append(x)
     return ','.join(output)
+
